+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # domino x.x.x (not yet released)
 * Allow attributes named 'xmlns' (#112)
+* Make DOMTokenList add/remove variadic (#109)
 
 # domino 2.0.0 ( 8 Nov 2017)
 * Fix potential O(N^2) slowdown in FilteredElementList#item.