+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# domino x.x.x (not yet released)
+
 # domino 2.0.1 (14 Feb 2018)
 * Allow attributes named 'xmlns' (#112)
 * Make DOMTokenList add/remove variadic (#109)