+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # domino x.x.x (not yet released)
 * Add TypeScript definitions (#103)
 * Add `flex` CSS styles (#119, #120)
+* Fix Element#matches with ~= selectors (#121)
 
 # domino 2.0.1 (14 Feb 2018)
 * Allow attributes named 'xmlns' (#112)